+++ src/main/java/com/picnee/travel/api/TokenController.java
@@ -3,7 +3,10 @@ package com.picnee.travel.api;
 import com.picnee.travel.api.in.TokenApi;
 import com.picnee.travel.domain.token.dto.CreateOauthToken;
 import com.picnee.travel.domain.token.service.TokenService;
+import com.picnee.travel.domain.user.dto.req.AuthenticatedUserReq;
+import com.picnee.travel.global.jwt.dto.res.AccessTokenRes;
 import com.picnee.travel.global.jwt.dto.res.JwtTokenRes;
+import com.picnee.travel.global.security.annotation.AuthenticatedUser;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
@@ -12,10 +15,7 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import static org.springframework.http.HttpStatus.*;
 
@@ -51,5 +51,24 @@ public class TokenController implements TokenApi {
         response.addHeader(HttpHeaders.SET_COOKIE, refreshTokenCookie.toString());
 
         return ResponseEntity.status(CREATED).build();
+    }
+
+    @PostMapping("/reissue")
+    public ResponseEntity<AccessTokenRes> reGenerateToken(@AuthenticatedUser AuthenticatedUserReq auth,
+                                                          @RequestHeader("RefreshToken") String refreshToken,
+                                                          HttpServletResponse response) {
+        AccessTokenRes res = tokenService.reissueToken(auth, refreshToken, response);
+
+        ResponseCookie accessTokenCookie = ResponseCookie.from("ACCESS_TOKEN", res.getAccessToken())
+                .httpOnly(true)
+                .secure(true)
+                .path("/")
+                .maxAge(24 * 60)
+                .sameSite("None")
+                .build();
+
+        response.addHeader(HttpHeaders.SET_COOKIE, accessTokenCookie.toString());
+
+        return ResponseEntity.status(OK).build();
     }
 }
